#12 Describe operations with type params
