+++ src/api/any-file.ts
@@ -1,12 +1,4 @@
-export interface AnyFile {
+import {LocalFile} from "./local-file";
+import {S3File} from "./s3-file";
 
-    /**
-     * The S3 bucket containing the file, undefined if it's a local file
-     */
-    bucket?: string;
-
-    /**
-     * The path to the local file or the S3 key
-     */
-    key: string;
-}
+export type AnyFile = LocalFile | S3File;
