+++ src/api/copy-request.ts
@@ -1,0 +1,7 @@
+import {WriteOptions} from "./write-options";
+
+export interface CopyRequest<T> {
+    source: T;
+    destination: T;
+    options: WriteOptions;
+}
