--- src/api/fs.error.ts
@@ -1,4 +1,0 @@
-export enum FsError {
-    LocalLink = "Can't get a link for a local file",
-    LocalFileWait = "Waiting for a local file to exist isn't supported yet"
-}
