+++ src/api/index.ts
@@ -1,8 +1,7 @@
 export * from "./any-file";
 export * from "./scanned-file";
-export * from "./fs.error";
-export * from "./complete";
 export * from "./progress-listener";
 export * from "./write-options";
 export * from "./s3-file";
 export * from "./scanned-s3-file";
+export * from "./copy-request";
