+++ src/api/local-file.ts
@@ -1,0 +1,3 @@
+export interface LocalFile {
+    key: string;
+}
