+++ src/api/scanned-s3-file.ts
@@ -1,4 +1,4 @@
-import {ScannedFile} from "./scanned-file";
-import {Complete} from "./complete";
+import {S3File} from "./s3-file";
+import {Scanned} from "./scanned";
 
-export type ScannedS3File = Complete<ScannedFile>;
+export type ScannedS3File = Scanned<S3File>;
