+++ src/s3/resolve-destination-request.ts
@@ -1,7 +1,8 @@
-import {AnyFile, ScannedFile} from "../api";
+import {LocalFile} from "../api/local-file";
+import {Scanned} from "../api/scanned";
 
-export interface ResolveDestinationRequest {
-    source: ScannedFile;
+export interface ResolveDestinationRequest<S extends LocalFile, D extends LocalFile> {
+    source: Scanned<S>;
     sourceFolder: string;
-    destinationFolder: AnyFile;
+    destinationFolder: D;
 }
