+++ src/web/web-file-service.ts
@@ -1,0 +1,5 @@
+import {S3FileService} from "../s3/s3-file-service";
+
+export class WebFileService extends S3FileService {
+
+}
